requirements.txt
win10toast
pillow
